Make RuleList property of Configuration Allows to access specified ruleList through current configuration
